+++ Source/PluginEditor.h
@@ -60,7 +60,7 @@ class FxseqAudioProcessorEditor  : public juce::AudioProcessorEditor,
         void loadPreset(std::string presetName);
 
         int greatestClockMult=4;
-        std::vector<std::string> fxNamesStr={"Chopper","Echo","Filter","Crusher","Distortion","Repeater","Chorus","Stretcher"};
+        std::vector<std::string> fxNamesStr={"Chopper","Echo","Filter","Crusher","Distortion","Repeater","RingMod","PitchShifter"};
         std::vector<int> clipboard;
 
         std::string currentSkin="default",currentPreset="default",rootPath="/home/pwner/dev/fxseq/Ressources/";
@@ -98,8 +98,8 @@ class FxseqAudioProcessorEditor  : public juce::AudioProcessorEditor,
                                           EffectComponent(3,this,"Crusher","combo4","slider4","programbutton4"),
                                           EffectComponent(4,this,"Distortion","combo5","slider5","programbutton5"),
                                           EffectComponent(5,this,"Repeater","combo6","slider6","programbutton6"),
-                                          EffectComponent(6,this,"Chorus","combo7","slider7","programbutton7"),
-                                          EffectComponent(7,this,"Stretcher","combo8","slider8","programbutton8"),
+                                          EffectComponent(6,this,"RingMod","combo7","slider7","programbutton7"),
+                                          EffectComponent(7,this,"PitchShifter","combo8","slider8","programbutton8"),
                                          };
 
         OptionsComponent options= OptionsComponent(this,"combo1","slider1","textbutton1");